backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.0
pydantic==2.9.2
pydantic-settings==2.6.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.17
cryptography==43.0.3
redis==5.2.0
celery==5.4.0
stripe==11.3.0
httpx==0.27.2
aiofiles==24.1.0
jinja2==3.1.4
websockets==14.1
psutil==6.1.0
